Move find file related functions from misc1.c, misc2.c and window.c to findfile.c
