+++ src/Make_dice.mak
@@ -46,6 +46,7 @@ SRC = \
 	ex_getln.c \
 	farsi.c \
 	fileio.c \
+	findfile.c \
 	fold.c \
 	getchar.c \
 	hardcopy.c \
@@ -105,6 +106,7 @@ OBJ =	o/arabic.o \
 	o/ex_getln.o \
 	o/farsi.o \
 	o/fileio.o \
+	o/findfile.o \
 	o/fold.o \
 	o/getchar.o \
 	o/hardcopy.o \
@@ -202,6 +204,8 @@ o/ex_getln.o:	ex_getln.c  $(SYMS)
 o/farsi.o:	farsi.c  $(SYMS)
 
 o/fileio.o:	fileio.c  $(SYMS)
+
+o/findfile.o:	findfile.c  $(SYMS)
 
 o/fold.o:	fold.c  $(SYMS)
